Added level 1 & 2 prefab

+++ Assets/Project/EnemyManagement/WaveInformation.cs
@@ -15,7 +15,7 @@ namespace Assets.Project.Scripts
         public int totalSpawns;
         [NonSerialized]
         public int spawnsLeft;
-        public int baseSpawnDelay = 1;
+        public float baseSpawnDelay = 1;
         public int pathId;
     }
 }
